Add enhanced React components for improved user experience
- Add FilterPanel.jsx for vulnerability filtering and search
- Add ProgressBar.jsx with detailed scanning progress tracking
- Add TokenManager.jsx for secure GitHub token management
- Add UploadArea.jsx for drag-and-drop file uploads
- Add VulnerabilityCard.jsx with expandable vulnerability details
- Create ScanContext.jsx for centralized state management
- Integrate modular scanner components into existing UI
