+++ src/components/VulnerabilityCard.jsx
@@ -1,0 +1,385 @@
+import React from 'react';
+import { patterns, patternCategories, recommendations } from '../lib/patterns';
+import { vulnerabilityGuides } from '../lib/proactiveControlsData';
+import { Shield } from 'lucide-react';
+
+/**
+ * Display line numbers with possible code snippet expansion.
+ */
+const FileLineNumbers = ({ vuln, file }) => {
+  const [expanded, setExpanded] = React.useState(false);
+  
+  // Get code lines for this specific file from the new allCodeLines structure
+  const codeLines = vuln.allCodeLines ? vuln.allCodeLines[file] : vuln.codeLines;
+  
+  // Show code lines for ALL scan types if available
+  if (codeLines && codeLines.length > 0) {
+    const visibleLines = expanded ? codeLines : codeLines.slice(0, 3);
+    return (
+      <div className="mt-2 space-y-2 bg-gray-800 p-3 rounded">
+        {visibleLines.map(({ line, code, isMinified, isHtml, matchedText }) => (
+          <div key={line} className="flex items-start space-x-2">
+            <span className="text-gray-500 select-none w-12 text-right font-mono">{line}</span>
+            <div className="flex-1">
+              {isHtml ? (
+                <pre
+                  className={`text-gray-300 overflow-x-auto font-mono text-sm whitespace-pre-wrap ${
+                    isMinified ? 'bg-gray-900/50 p-2 rounded' : ''
+                  }`}
+                  dangerouslySetInnerHTML={{ __html: code }}
+                />
+              ) : (
+                <pre className="text-gray-300 overflow-x-auto font-mono text-sm whitespace-pre-wrap">
+                  {code.replace(/\*\*(.*?)\*\*/g, '→ $1 ←')}
+                </pre>
+              )}
+              {matchedText && (
+                <div className="text-xs text-yellow-400 mt-1">
+                  Suspicious pattern: <code className="bg-gray-900 px-1 rounded">{matchedText}</code>
+                </div>
+              )}
+            </div>
+          </div>
+        ))}
+        {!expanded && codeLines.length > 3 && (
+          <button
+            onClick={() => setExpanded(true)}
+            className="text-blue-400 text-xs hover:underline mt-2"
+          >
+            Show {codeLines.length - 3} more lines
+          </button>
+        )}
+      </div>
+    );
+  }
+
+  // Fallback for old format - show line numbers only
+  const lines = vuln.allLineNumbers[file];
+  if (!lines || lines.length === 0) return null;
+  
+  return (
+    <div className="mt-2">
+      <span className="text-gray-300">Lines: {lines.join(', ')}</span>
+    </div>
+  );
+};
+
+/**
+ * Vulnerability Card - toggles between normal "CVE details" and "Protection Guide" 
+ */
+const VulnerabilityCard = ({ vulnerability, isExpanded, onToggleExpand, cardId }) => {
+  const [isGuideView, setIsGuideView] = React.useState(false);
+
+  // Use vulnerability prop instead of vuln
+  const vuln = vulnerability;
+
+  // Mapping raw vulnerability type strings to the keys used in proactiveControlsData.
+  const vulnerabilityGuideKeyMap = {
+    'a09:2021 - security logging and monitoring failures': 'securityLogging',
+    'insufficientlogging': 'securityLogging',
+    'inadequatelogging': 'securityLogging',
+    'securitylogging': 'securityLogging',
+    'commandinjection': 'commandExecution',
+    'commandexecution': 'commandExecution',
+    'insecuresubmission': 'insecureSubmission',
+    'insecuretransmission': 'insecureSubmission',
+    'nosqlinjection': 'noSqlInjection',
+    'hardcodedsecret': 'hardcodedSecret',
+    'weakcrypto': 'weakCrypto',
+    'insecurecryptousage': 'insecureCryptoUsage',
+    'openredirect': 'openRedirect',
+    'pathtraversal': 'pathTraversal',
+  };
+  const baseKey = vuln.type.toLowerCase().replace(/\s+/g, '');
+  const normalizedType = vulnerabilityGuideKeyMap[baseKey] || vuln.type;
+  const rec = recommendations[normalizedType] || recommendations[vuln.type];
+  const matchedPattern = patterns[normalizedType] ? patterns[normalizedType].pattern.toString() : '';
+  
+  // Red/Blue Team data
+  const guideData = vulnerabilityGuides[normalizedType] || {};
+
+  // Severity styling
+  const severityBadge = {
+    CRITICAL: 'bg-red-500 text-white',
+    HIGH: 'bg-orange-500 text-white',
+    MEDIUM: 'bg-yellow-500 text-gray-800',
+    LOW: 'bg-blue-500 text-white'
+  }[vuln.severity] || 'bg-gray-500 text-white';
+
+  // Toggle top-right button
+  const handleGuideToggle = (e) => {
+    e.stopPropagation();
+    setIsGuideView(!isGuideView);
+  };
+
+  return (
+    <div className="border border-gray-700 rounded-lg shadow-sm text-sm">
+      {/* Header */}
+      <div
+        className="w-full flex items-center justify-between bg-gray-800 hover:bg-gray-700 transition-colors cursor-pointer p-4"
+        onClick={() => onToggleExpand(cardId)}
+      >
+        <div>
+          <div className="flex items-center space-x-2">
+            <span className={`text-xs font-semibold py-1 px-2 rounded-full uppercase ${severityBadge}`}>
+              {vuln.severity}
+            </span>
+            <span className="text-xs text-gray-400">
+              Found in {vuln.files.length} file{vuln.files.length !== 1 ? 's' : ''}
+            </span>
+          </div>
+          <h3 className="text-lg font-medium mt-1">{vuln.description}</h3>
+        </div>
+        <div className="flex items-center space-x-2">
+          <button
+            onClick={(e) => { e.stopPropagation(); handleGuideToggle(e); }}
+            className="h-8 px-3 bg-blue-600 text-white rounded hover:bg-blue-700 transition-colors text-xs flex items-center gap-1 min-w-[120px] justify-center"
+          >
+            <Shield className="w-4 h-4" />
+            {isGuideView ? "CVE Details" : "Protection Guide"}
+          </button>
+          <svg
+            className={`w-5 h-5 transform transition-transform ${isExpanded ? 'rotate-180' : ''}`}
+            viewBox="0 0 24 24"
+            fill="none"
+            stroke="currentColor"
+            strokeWidth="2"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+          >
+            <polyline points="6 9 12 15 18 9"></polyline>
+          </svg>
+        </div>
+      </div>
+      {/* Expanded content */}
+      {isExpanded && (
+        <div className="p-4 bg-gray-700 text-gray-200">
+          {!isGuideView ? (
+            /* =============== CVE Details View =============== */
+            <div>
+              {/* File list */}
+              <div className="files-list mb-4 text-sm text-gray-300">
+                {vuln.files.length > 0 ? (
+                  <div>Found in {vuln.files.length} file{vuln.files.length > 1 ? 's' : ''}:</div>
+                ) : (
+                  <div>No files recorded.</div>
+                )}
+                {vuln.files.map((file, idx) => (
+                  <details
+                    key={`${file}-${idx}`}
+                    className="file-item border border-gray-600 rounded-md mt-2"
+                  >
+                    <summary className="px-3 py-2 bg-gray-800 rounded-t-md cursor-pointer hover:bg-gray-700">
+                      {file}
+                    </summary>
+                    <div className="p-3 bg-gray-700 rounded-b-md">
+                      <FileLineNumbers vuln={vuln} file={file} />
+                    </div>
+                  </details>
+                ))}
+              </div>
+              {/* Recommendation Section */}
+              {rec ? (
+                <div className="bg-gray-900 rounded-lg p-4 border border-gray-700">
+                  {/* Why it Matters */}
+                  <div className="mb-4">
+                    <h4 className="text-sm font-semibold text-blue-300 mb-2">Why it Matters</h4>
+                    <p className="text-gray-300 text-sm">
+                      Command injection can allow attackers to execute arbitrary system commands.
+                    </p>
+                  </div>
+
+                  {/* What to Do */}
+                  <div>
+                    <h4 className="text-sm font-semibold text-blue-300 mb-2">What to Do</h4>
+                    <div className="space-y-4 text-sm text-gray-300">
+                      {rec.recommendation.split('<div class="example-block">').map((section, index) => {
+                        if (section.includes('example-label')) {
+                          // This is a code example section
+                          const isVulnerable = section.includes('✕ Vulnerable:');
+                          const label = isVulnerable ? '✕ Vulnerable:' : '✓ Safe:';
+                          const code = section.split('<code>')[1]?.split('</code>')[0];
+                          
+                          return code ? (
+                            <div key={index} className="mt-4">
+                              <div className="flex items-center gap-2 mb-1">
+                                <span className={`${isVulnerable ? 'text-red-400' : 'text-green-400'} text-sm`}>
+                                  {label}
+                                </span>
+                              </div>
+                              <pre className="bg-gray-950 p-3 text-gray-300 font-mono rounded border border-gray-800 overflow-x-auto whitespace-pre-wrap break-words">
+                                <code>{code.trim()}</code>
+                              </pre>
+                            </div>
+                          ) : null;
+                        }
+
+                        // Regular text and numbered items
+                        return section.split('\n').map((line, lineIndex) => {
+                          if (!line.trim()) return null;
+                          
+                          // If it's a numbered item
+                          if (line.match(/^\d+\./)) {
+                            return (
+                              <div key={`${index}-${lineIndex}`} className="flex items-start">
+                                <span className="mr-2">{line.match(/^\d+\./)[0]}</span>
+                                <span>{line.replace(/^\d+\./, '').trim()}</span>
+                              </div>
+                            );
+                          }
+                          
+                          // Regular text
+                          return <p key={`${index}-${lineIndex}`}>{line.trim()}</p>;
+                        });
+                      })}
+                    </div>
+                  </div>
+
+                  {/* Code Examples */}
+                  {rec.examples && (
+                    <div className="mt-4">
+                      {rec.examples.map((example, index) => (
+                        <div key={index} className="mt-4">
+                          <div className="flex items-center gap-2 mb-1">
+                            {example.type === 'vulnerable' ? (
+                              <span className="text-red-400 text-sm">✕ Vulnerable:</span>
+                            ) : (
+                              <span className="text-green-400 text-sm">✓ Safe:</span>
+                            )}
+                          </div>
+                          <pre className="bg-gray-950 p-3 text-sm text-gray-300 font-mono rounded border border-gray-800 overflow-x-auto whitespace-pre-wrap break-words">
+                            <code>{example.code}</code>
+                          </pre>
+                        </div>
+                      ))}
+                    </div>
+                  )}
+
+                  {/* References */}
+                  {rec.references && rec.references.length > 0 && (
+                    <div className="mt-6 pt-4 border-t border-gray-700">
+                      <h4 className="text-sm font-semibold text-blue-300 mb-2">References</h4>
+                      <ul className="space-y-1">
+                        {rec.references.map((ref, i) => (
+                          <li key={i}>
+                            <a
+                              href={ref.url}
+                              className="text-sm text-blue-400 hover:text-blue-300 hover:underline inline-flex items-center gap-1"
+                              target="_blank"
+                              rel="noreferrer"
+                            >
+                              {ref.title}
+                              <svg className="w-3 h-3" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
+                                <path d="M18 13v6a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V8a2 2 0 0 1 2-2h6" />
+                                <polyline points="15 3 21 3 21 9" />
+                                <line x1="10" y1="14" x2="21" y2="3" />
+                              </svg>
+                            </a>
+                          </li>
+                        ))}
+                      </ul>
+                    </div>
+                  )}
+
+                  {/* Detection Pattern */}
+                  {matchedPattern && (
+                    <div className="mt-6 pt-4 border-t border-gray-700">
+                      <h4 className="text-sm font-semibold text-blue-300 mb-2">Detection Pattern</h4>
+                      <pre className="bg-gray-950 p-3 text-sm text-gray-300 font-mono rounded border border-gray-800 overflow-x-auto whitespace-pre-wrap break-words">
+                        <code>{matchedPattern}</code>
+                      </pre>
+                      {(vuln.category || vuln.subcategory) && (
+                        <div className="mt-2 text-sm text-gray-400">
+                          <div>Category: {Object.keys(patternCategories).find(k => patternCategories[k] === vuln.category)} ({vuln.category})</div>
+                          <div>Subcategory: {vuln.subcategory}</div>
+                        </div>
+                      )}
+                    </div>
+                  )}
+                </div>
+              ) : (
+                <div className="bg-gray-900 rounded-lg p-4 border border-gray-700 text-gray-300">
+                  No recommendation found for "{vuln.type}".
+                </div>
+              )}
+            </div>
+          ) : (
+            /* =============== Protection Guide (Red/Blue) View =============== */
+            <div className="space-y-6">
+              {/* Try a Simulation (Coming Soon) */}
+              <div className="border-2 border-dashed border-gray-600 rounded-lg p-4 text-center mb-4 bg-gray-800">
+                <span className="text-blue-400 font-semibold">Try a Simulation</span>
+                <span className="ml-2 text-sm text-yellow-400">(Coming Soon!)</span>
+              </div>
+              {/* Main text content */}
+              {guideData.title && (
+                <div className="bg-gray-800/50 rounded-lg p-4 border border-gray-600">
+                  <h2 className="text-lg font-semibold text-blue-300 mb-3">{guideData.title}</h2>
+                  <div
+                    className="prose prose-invert text-gray-100 max-w-none"
+                    dangerouslySetInnerHTML={{ __html: guideData.content || '' }}
+                  />
+                </div>
+              )}
+
+              {/* Red Team */}
+              {guideData.redTeam && (
+                <div className="bg-gray-800/50 rounded-lg p-4 border border-red-500/40">
+                  <h3 className="text-md font-semibold text-red-400 mb-2">Red Team</h3>
+                  <div
+                    className="prose prose-invert text-gray-100 max-w-none"
+                    dangerouslySetInnerHTML={{ __html: guideData.redTeam }}
+                  />
+                </div>
+              )}
+
+              {/* Blue Team - check if separate OS sections or single fallback */}
+              {guideData.blueTeamWindows || guideData.blueTeamMac || guideData.blueTeamLinux ? (
+                <div className="space-y-3">
+                  {guideData.blueTeamWindows && (
+                    <div className="bg-gray-800/50 rounded-lg p-4 border border-blue-500/30">
+                      <div
+                        className="prose prose-invert text-gray-100 max-w-none"
+                        dangerouslySetInnerHTML={{ __html: guideData.blueTeamWindows }}
+                      />
+                    </div>
+                  )}
+                  {guideData.blueTeamMac && (
+                    <div className="bg-gray-800/50 rounded-lg p-4 border border-blue-500/30">
+                      <div
+                        className="prose prose-invert text-gray-100 max-w-none"
+                        dangerouslySetInnerHTML={{ __html: guideData.blueTeamMac }}
+                      />
+                    </div>
+                  )}
+                  {guideData.blueTeamLinux && (
+                    <div className="bg-gray-800/50 rounded-lg p-4 border border-blue-500/30">
+                      <div
+                        className="prose prose-invert text-gray-100 max-w-none"
+                        dangerouslySetInnerHTML={{ __html: guideData.blueTeamLinux }}
+                      />
+                    </div>
+                  )}
+                </div>
+              ) : guideData.blueTeam ? (
+                <div className="bg-gray-800/50 rounded-lg p-4 border border-blue-500/40">
+                  <h3 className="text-md font-semibold text-blue-400 mb-2">Blue Team</h3>
+                  <div
+                    className="prose prose-invert text-gray-100 max-w-none"
+                    dangerouslySetInnerHTML={{ __html: guideData.blueTeam }}
+                  />
+                </div>
+              ) : (
+                <div className="bg-gray-800/50 rounded-lg p-4 border border-gray-600">
+                  <p className="text-gray-100">No Blue Team guidance found for this vulnerability.</p>
+                </div>
+              )}
+            </div>
+          )}
+        </div>
+      )}
+    </div>
+  );
+};
+
+export default VulnerabilityCard;
